fix: resolve the second round of review comments on #495
Four findings addressed, two declined with reasons.

merge-windows-updater.py: PYTHONIOENCODING only covers stdin/stdout/stderr,
not Path.read_text/write_text, which fall back to the locale encoding -
cp1252 on the Windows runner this script runs on. My #500 fix pinned the
console encoding and left the file I/O exposed, so a single non-ASCII
character in latest.json would still have failed the read. Encoding is now
pinned on all three file operations.

LlmProviderPicker.tsx: InstallOutcome, ProviderStatus, ProviderTestResult
and ProviderTestOutcome are Rust-mirrored invoke contracts and belong in
ui/lib/api-types.ts per the repo rule, so they cannot drift from
src/llm/detect.rs unnoticed. All four moved, not just the one flagged -
splitting them would leave exactly the divergence the rule prevents. The
picker re-exports them so existing import sites keep resolving.

src/llm/detect.rs: install_provider and cursor_sign_in shell out to vendor
installers - real work, and the observability guideline requires a span with
an explicit failure status. Both are now instrumented, recording ok and the
resolved cli_path. install_failed is the single funnel every failure path in
both functions returns through, so recording on the current span there marks
whichever is running as failed without threading a handle through a dozen
early returns.

ui/__tests__: sourceOf and stripComments were near-identical
comment-stripping helpers in two suites. Extracted to
__tests__/helpers/source.ts.

Not taken:

  * Pinning dtolnay/rust-toolchain@master to a SHA. It appears in all seven
    places across every workflow; pinning one file is inconsistent rather
    than safer. The review cites a repo policy, but nothing in
    .github/workflows is pinned today. Belongs in its own repo-wide change.
  * Loosening the placeholder heuristic. value_is_set rejects anything
    containing `your-`, `_your_` or `-here`, which can in principle reject a
    valid Jira URL, but the failing examples are contrived and tightening it
    is a product judgement about what a real credential looks like.
